- Finalisation Menu - Ajout Catégorie - Début ajout photo non finalisé

+++ src/main/java/isen/java_contact_app/view/MainLayoutController.java
@@ -1,23 +1,22 @@
 package isen.java_contact_app.view;
 
-import java.util.Observable;
-
 import isen.java_contact_app.model.Person;
 import isen.java_contact_app.service.PersonService;
-import javafx.collections.ObservableList;
-import javafx.application.Application;
 import javafx.fxml.FXML;
 
 import java.io.File;
 import java.io.IOException;
 import java.time.LocalDate;
+import java.util.Optional;
 
 import isen.java_contact_app.service.StageService;
+import isen.java_contact_app.service.UserService;
 import isen.java_contact_app.service.ViewService;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.SeparatorMenuItem;
 import javafx.stage.DirectoryChooser;
-import javafx.stage.Stage;
 
 public class MainLayoutController {
 	
@@ -40,6 +39,7 @@ public class MainLayoutController {
 	}
 
 	public void changeUser() {
+		UserService.setCurrentUser(null);
 		StageService.showView(ViewService.getView("HomeScreen"));
 	}
 
@@ -47,9 +47,8 @@ public class MainLayoutController {
 	{
 		System.out.println("exportation des données");
 		File directory = dataFolder("export");
-		System.out.println(directory);
 		if (directory != null) {
-			pathFolder = directory.getPath();
+			UserService.getCurrentUser().setPathFolderContact(directory.getPath());
 			File newDirectory = new File(directory, LocalDate.now().toString());
 			int numberInstanceDirectory = 2;
 			while (newDirectory.exists()) {
@@ -69,31 +68,42 @@ public class MainLayoutController {
 	
 	public void importData() throws IOException{
 		System.out.println("importation des données");
-		File directory = dataFolder("import");
-		if (directory != null) {
-			pathFolder = directory.getParent();
-			File[] files = directory.listFiles();		// Contient le chemin absolu de chaque éléments du dossier choisi précédemment
-			PersonService.deleteListPerson();
-			for (int i = 0; i < files.length; i++) {
-				if (files[i].getAbsolutePath().endsWith("vcf")) {
-					PersonService.addPerson(Person.importFile(files[i]));
+		Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+		alert.initOwner(StageService.getPrimaryStage());
+		alert.setTitle("INFORMATION");
+		alert.setHeaderText("Don't forget to make a backup/export of the current database," + "\n" + "Make an import will permanently replace it." + "\n" + "Continue?");
+		Optional<ButtonType> option = alert.showAndWait();
+		if (option.get() == ButtonType.OK) {
+			File directory = dataFolder("import");
+			if (directory != null) {
+				UserService.getCurrentUser().setPathFolderContact(directory.getParent());
+				File[] files = directory.listFiles();
+				PersonService.deleteListPerson();
+				for (int i = 0; i < files.length; i++) {
+					if (files[i].getAbsolutePath().endsWith("vcf")) {
+						PersonService.addPerson(Person.importFile(files[i]));
+					}
 				}
 			}
-		}
+    	}
 	}
 	
 	
 	public File dataFolder(String action) {
 		final DirectoryChooser directoryChooser = new DirectoryChooser();
 		directoryChooser.setTitle("Select the contact's folder to " + action);
-    	directoryChooser.setInitialDirectory(new File(pathFolder));
-		File directory;
+		directoryChooser.setInitialDirectory(new File(UserService.getCurrentUser().getPathFolderContact()));
+		File directory = null;
         try{
-        	directory = directoryChooser.showDialog(StageService.getPrimaryStage());
-        }catch(IllegalArgumentException e) {
-        	directoryChooser.setInitialDirectory(new File(System.getProperty("user.home")));
-        	directory = directoryChooser.showDialog(StageService.getPrimaryStage());
-        }
+        	try{
+	        	directory = directoryChooser.showDialog(StageService.getPrimaryStage());
+	        }catch(IllegalArgumentException e){
+				directoryChooser.setInitialDirectory(new File(System.getProperty("user.home")));
+	    		directory = directoryChooser.showDialog(StageService.getPrimaryStage());
+	        }
+	    }catch (Exception e) {
+	    	e.printStackTrace();
+	    }
         return directory;
 	}
 	
@@ -114,11 +124,6 @@ public class MainLayoutController {
 		exportDataMenuItem.setVisible(visible);
 		changeUserMenuItem.setVisible(visible);
 		separatorMenuItem.setVisible(visible);
-	}
-	
-	@FXML
-	private void initialize() {
-		pathFolder = System.getProperty("user.home");
 	}
 
 }
